layout/generic/test/test_bug589623.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=589623
-->
<head>
  <title>Test for Bug 589623</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=589623">Mozilla Bug 589623</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script>
/** Test for Bug 589623 **/
var sel = getSelection();
sel.removeAllRanges();
function t(m) {
  try {
    sel[m]();
    ok(false, "Should not be here");
  } catch(e) {
    is(e.name, "InvalidStateError", "Should be an InvalidStateError");
    ok(e instanceof DOMException, "Should be a DOMException");
    is(e.code, DOMException.INVALID_STATE_ERR, "Should be an INVALID_STATE_ERR");
  }
}
t("collapseToStart");
t("collapseToEnd");
</script>
</pre>
</body>
</html>
